user-exporter/app/requirements.txt
mysql-connector-python==8.0.31
google-cloud-logging==3.2.5
